+++ gen/doc.go
@@ -1,0 +1,2 @@
+// Package gen contains some useful generator functions like Range, Take, and Repeat
+package gen